analytics_service/requirements.txt
mysql-connector-python
pymongo
numpy
